Section 2 - Units 4 to 6
Uploaded the art assets we will be using for the course. Then we laid out the sprites and tiles into the project. From there, we created a tile mapper script that allows us to interact with our tile prefabs. So far we have randomized the tiles graphics as well as created a mouse hover animation to play