Section 2 - Units 9 - 10
We made big changes and updates here. Fixing up our tile-movement selection and ensuring the tiles were rest with a GameManager funciton known as Reset Tiles. From there, we went to work on making a move function for the units to move to the desired tiles. This called two while loops inn an enum that let the player move to the tile position.