Created a New Scene, Fixing up our Prefabs and Objects
In this push, I fixed up the scene by making a new one. Resdigning it, making everything look nice, the works. I wanted to fix the bug where the Unit's Z pos was conflicting with the z-pose of the tiles and I managed to do so!

I also added a fun animation for the movement of the units!